Remove Temporal bridge module
