+++ src/temporal/workflows/monthly_post.py
@@ -6,7 +6,6 @@ from temporalio import workflow
 
 from ..activities import notifications as notification_activities
 from ..activities import submissions as submission_activities
-from ..activities import temporal_bridge as bridge_activities
 from ..shared import (
     CreateMonthlyPostInput,
     REDDIT_RETRY_POLICY_CONSERVATIVE as REDDIT_RETRY_POLICY,
@@ -44,20 +43,12 @@ class MonthlyPostWorkflow:
         )
 
         # Discover the current submission ID (the one we're replacing).
-        subs = await workflow.execute_activity(
-            bridge_activities.query_polling_submissions,
-            start_to_close_timeout=timedelta(seconds=30),
+        active_submissions = await workflow.execute_activity(
+            submission_activities.fetch_active_submission_ids,
+            start_to_close_timeout=timedelta(seconds=60),
+            retry_policy=REDDIT_RETRY_POLICY,
         )
-        old_submission_id = subs.current_submission_id
-
-        # Fallback: if polling workflow wasn't reachable, bootstrap from Reddit.
-        if old_submission_id is None:
-            fallback = await workflow.execute_activity(
-                submission_activities.fetch_active_submission_ids,
-                start_to_close_timeout=timedelta(seconds=60),
-                retry_policy=REDDIT_RETRY_POLICY,
-            )
-            old_submission_id = fallback.current_submission_id
+        old_submission_id = active_submissions.current_submission_id
 
         # Create the new monthly post (idempotent).
         new_submission_id = await workflow.execute_activity(
